+++ insts/std/LPCPLAY/buzz.c
@@ -1,6 +1,6 @@
 #include "ugens.h"
 
-float buzz(float amp, float si, float hn, float *f, float *phs)
+float buzz(float amp, float si, float hn, double *f, float *phs)
 {
 	int j,k;
 	float q,d,h2n,h2np1;
@@ -25,7 +25,7 @@ float buzz(float amp, float si, float hn, float *f, float *phs)
 }
 
 float *
-bbuzz(float amp, float si, float hn, float *f, float *phs, float *out, long alen)
+bbuzz(float amp, float si, float hn, double *f, float *phs, float *out, long alen)
 {
 	int i,j,k;
 	float q,d;
